src/configs/requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
python-dotenv
pydantic[email]
passlib
bcrypt
python-multipart
python-jose
jinja2
alembic
aiofiles